requirements.txt
mss
pyautogui
pillow
pynput
openai
textual
